CORE-134: In Core Ethereum, allow GETH/LES requests (BlockHeader, BlockBodies,…) to have unlimited numbers of results - rather than being limited to GETH per-request limits.  Use a ‘Provisioner’ to collect individual message, within GETH per-reqeust limits, to construct the ‘unlimited’ result.  Structure for Parity/PIP.

+++ ethereum/ewm/BREthereumEWM.c
@@ -528,7 +528,7 @@ ewmGetWallets (BREthereumEWM ewm) {
 
 extern unsigned int
 ewmGetWalletsCount (BREthereumEWM ewm) {
-    return array_count(ewm->wallets);
+    return (unsigned int) array_count(ewm->wallets);
 }
 
 extern BREthereumWalletId
